pyproject.toml
[project]
name = "iotawatt-sql"
version = "0.1.0"
description = "IoTaWatt SQL monitoring system with TimescaleDB, PostgREST, and Python utilities"
requires-python = ">=3.12"
dependencies = [
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "psycopg2-binary>=2.9.9",
    "typer>=0.20.0",
]


[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

